requirements.txt
# Requirements for Ghost Scanner
python-nmap==0.7.1
scapy==2.5.0
customtkinter==5.2.1